+++ Classix/OStreamDisassemblyWriter.h
@@ -34,8 +34,8 @@ public:
 	OStreamDisassemblyWriter(std::ostream& into);
 	
 	virtual void EnterSection(const PEF::InstantiableSection& section, uint32_t sectionIndex) override;
-	virtual void EnterLabel(const PPCVM::Disassembly::InstructionRange& label, intptr_t labelAddress) override;
-	virtual void VisitOpcode(const PPCVM::Disassembly::DisassembledOpcode& opcode, intptr_t opcodeAddress, const std::string* metadata) override;
+	virtual void EnterLabel(const PPCVM::Disassembly::InstructionRange& label, uint32_t labelAddress) override;
+	virtual void VisitOpcode(const PPCVM::Disassembly::DisassembledOpcode& opcode, uint32_t opcodeAddress, const std::string* metadata) override;
 	
 	virtual ~OStreamDisassemblyWriter() override;
 };
